IAccessible2: Evaluate "text-indent" attr for first line indent (#17638)
Fixes #13052

Summary of the issue:
NVDA did not announce the first line indent for paragraphs in LibreOffice Writer.

Description of user facing changes
NVDA supports the "text-indent" IAccessible2 object attribute. As a consequence, announcement of first line indent is now supported for LibreOffice 25.8 and later.

Description of development approach
Implement support for the "text-indent" object attribute as specified in the
IAccessible2 object attribute specification, which references the CSS spec that explains:

This property specifies the indentation of the first line
of text in a block container.

https://gerrit.libreoffice.org/c/core/+/180574 implements support in LibreOffice, i.e. reporting of that attribute by LibreOffice.

This commit implements evaluation of that IAccessible2 object attribute by NVDA.

As recommended in the IAccessible2 object attribute spec, LibreOffice reports the indentation in mm. Add a helper method that converts that to either cm or inches, depending on whether use of imperial measurements is enabled.