Remove contributors file from NVDA, read license as HTML (#17600)
Closes #17529
Closes #16922

Summary of the issue:
There is a transient bug with Windows that prevents opening the "contributors" and "license" file in notepad.
The contributors file is no longer maintained or useful, so this should instead be archived.
The license file is written in markdown, but displayed as plain text in notepad, or as a plain text scrollable text window. This makes it hard for users to read.

Description of user facing changes
The license file is now displayed in HTML as a browseable message, generated from markdown.
From the installer, there is now a button to open the license as a message, rather than a direct embedded panel
The installer now has a more thorough agreement message to confirm the user has read the license
The contributors file can no longer be accessed from NVDA
Description of development approach
Use markdown module to convert markdown to html
Use nh3 to sanitize generated HTML
Generate and cache HTML from the license only when required, on first load.