Add ability to customize automatic update channels for add-ons (#17597)
Part of #3208
Closes #15261

Summary of the issue:
Add-ons currently only automatically update to the same channel with automatic updates enabled
This is quite restrictive as many people may want the following scenarios:

disabling specific add-ons from update, with automatic updates enabled by default
enabling only specific add-ons from update, disabling automatic updates by default
sticking to the latest version regardless of channel
migrating back to a stable version from a beta/dev version
sticking to the latest beta or dev version only
migrating from an externally installed version
Description of user facing changes
Added an "update channel" submenu for add-ons on the installed/update tabs of the add-on store.
This allows selecting a specific set of update channels on a per add-on basis.

Added an item to the add-on store settings panel for selecting the default update channel for all add-ons

Description of development approach
Created an add-on store settings cache, which also stores add-on specific information.
Migrated the "showWarning" setting to add-on store specific settings, so it can be saved independently of config profiles