Fix leading silence removal feature (#17699)
None. Fixes problems introduced by PR #17648.

Summary of the issue:
The leading silence trimming feature should only trim the silence part of the audio. However, a longer part of the speech audio ended up being chopped off, which makes distinguishing the consonants at the beginning more difficult.

In addition, trimming is incorrectly applied to non-speech audio, and applied even when the feature is turned off in Settings.

Description of user facing changes
The problem mentioned above should be fixed.

Description of development approach
Added missing check for _enableTrimmingLeadingSilence in _idleCheck.

Before sending the trimmed audio to the device, one silent sample will be inserted at the beginning. For some reason, if the first sample is not zero, the audio may be chopped off at the beginning. Although the exact cause is still unknown, prepending one silent sample seems to fix this issue. See this discussion.