Add `WS_EX_TOOLWINDOW` window style (#17711)
Fixed: #17696

Summary of the issue:
When the highlighter is enabled, the NVDA Highlighter Window icon is fixed on the taskbar after restarting Explorer.

Description of user facing changes
The NVDA Highlighter Window icon is no longer fixed in the taskbar after restarting Explorer.

Description of development approach
Add a WS_EX_TOOLWINDOW window style to the highlight window.