MS Word: The result of Word commands to expand or collapse a heading is now reported (#17545)
Related to #17499, but independent of it.

Summary of the issue:
In MS Word, control+alt+plus and control+alt+- allow to expand and collapse the content below a heading. NVDA does not report the result of this action.

Description of user facing changes:
NVDA now reports the result of the Word commands to expand and collapse headings.
This command does not exist in Draft mode, so nothing is reported when in this mode.
The command also does not report anything in Outline mode (limitation of this PR).

Description of development approach:
Use Word object model (if available) as done for many Word gestures, no matter if UIA is enabled or not

Testing strategy:
Tested with and without UIA.

Known issues with pull request::
This PR does not cover outline mode. The collapse / extend commands still don't report anything in Outline mode; see related issue #17645.