Fix prevent display from turning off (#17714)
Fixes #17705 
Fixup of #17651

Summary of the issue:
Pr #17651 introduced a feature flag in the general section of the config, but this is not supported yet.

Description of user facing changes
General settings panel can be reopened after saving.

Description of development approach
Rename the flag and make it a boolean.