Download without duplicates to override incompatibilities. (#17717)
Fixed #17655

Summary of the issue:
When there is an error downloading an Add-on that overrides an incompatibility, retrying the download will encounter an error. The reason for this is that the exception thrown when overriding incompatibilities is not handled for duplicates.

Description of user facing changes
The download can now be retried to override the incompatible Add-on.

No changelog is required for this PR as the Retry Download Add-on was introduced in 2025.1.

Description of development approach
Determine if an incompatibility has been overridden before overriding it.