Manually refresh the recognized content with `NVDA+f5` in a recognition result (#17715)
Summary of the issue:
Even if not using the automatic refresh of the content recognition, the user may want to update the recognized content once manually. Today, they needs to close the recog result and perform a recognition again.

Description of user facing changes
In a result of content recognition, the user can now press NVDA+f5 to refresh the result of recognized content. This refresh gesture is the same as the one to refresh the content of virtual buffers.

Description of development approach
In RefreshableRecogResultNVDAObject, added a script to refresh the recognition result. If automatic recognition is enabled, the script reports an error message.