Fix-up #17720 for sheets containing comma (#17754)
Fix-up of #17720

Summary of the issue:
#17720 fixed Excel element list dialog on non-English system, using the local list separator to request a range. Though, the way it was implemented could fail in case the Excel file or sheet names also contained a comma.

Description of user facing changes
Listing comments or formulas in Excel no longer fails when the name of the file or the sheet contains a comma.

Description of development approach
Split the range string on the last "!" character to isolate the cells range from the file/sheet names.
Replace the separator only in the cells range part.
Testing strategy: