Implement an extension point deprecation strategy (#17644)
Follow up of #17011
Slightly related to #17635, as this oversight was found there

Summary of the issue:
In #17011, `braille.filter_displaySize` was deprecated, but there was not yet a mechanism in place that logs warnings about usage of the deprecated extension point.

Description of user facing changes
Warnings in the log when `braille.filter_displaySize.register` is called.

Description of development approach
1. Added an optional keyword argument `_deprecationmessage` to `HandlerRegistrar`. If set, it will be logged when a handler is registered.
2. Removed remaining handlers in core, changed them into handlers for `filter_displayDimensions`
3. Unit test expansion, missing tests for `filter_displayDimensions`

Testing strategy:
Tested that a warning is logged when registering 

Known issues with pull request:
None known