GUI message when the configuration cannot be saved (#17827)
Summary of the issue:
When trying to save a string containing both `"""` and `'''`, it fails
due to configobj not supporting this. But the user is not warned, only a
warning and an error are logged.

Example:
* Open advanced settings
* In Regular expression for text paragraph navigation, add `'''"""` at
the end of the regexp.
* Press OK
* Press NVDA+control+C to save the config.

### Description of user facing changes
When the user manually saves the config (`NVDA+control+C` or NVDA menu
-> Save configuration), an error message error is displayed in case
saving the config fails.

### Description of development approach
* In addition to `PermissionError` which correspond to a specific use
case, catch any `Exception` for other unexpected use case and display
the appropriate message. For now, the root issue does not need to be
fixed since it does not correspond to a real use case, i.e. using `"""`
and `'''` in the paragraph nav regexp is not expected in real life.
* No change log for this fix which does not apply to a real-life use
case from what I know.

Testing strategy:
Manual test of the example described in the "Summary of the issue" of
this PR's description.

Known issues with pull request:
None